requirements.txt
numpy
panda
nltk
google
colorama
argparse